requirements.txt
# Web Framework
fastapi==0.135.3
uvicorn==0.44.0
starlette==1.0.0

# AI / LLM
langchain==1.2.15
langchain-core==1.2.27
langchain-groq==1.1.2
langchain-text-splitters==1.1.1
langgraph==1.1.6
groq==0.37.1

# Memory
mem0ai==1.0.11

# Vector Store / RAG
chromadb==1.5.6
qdrant-client==1.17.1

# Embeddings
sentence-transformers==5.4.0
torch==2.11.0
transformers==5.5.3
huggingface_hub==1.9.1

# Settings & Validation
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2

# Email Validation
email-validator==2.3.0
dnspython==2.8.0

# HTTP client
httpx==0.28.1
requests==2.33.1

# Database
SQLAlchemy==2.0.49

# Dashboard (Streamlit UI)
streamlit==1.56.0
pandas==2.3.3
altair==6.0.0

# Testing
pytest==9.0.3

# Utilities
numpy==2.2.6
tenacity==9.1.4
